# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Status": {
				"enum": [
					"ACTIVE",
					"INACTIVE",
					"DEACTIVATED"
				],
				"type": "string"
			},
...
			"CreateRouteDTO": {
				"properties": {
					"collegeId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"stops": {
						"items": {
							"properties": {
								"estimatedTime": {
									"type": "string"
								},
								"sequence": {
									"type": "number",
									"format": "double"
								},
								"longitude": {
									"type": "number",
									"format": "double"
								},
								"latitude": {
									"type": "number",
									"format": "double"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"longitude",
								"latitude",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"collegeId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRouteDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"stops": {
						"items": {
							"properties": {
								"estimatedTime": {
									"type": "string"
								},
								"sequence": {
									"type": "number",
									"format": "double"
								},
								"longitude": {
									"type": "number",
									"format": "double"
								},
								"latitude": {
									"type": "number",
									"format": "double"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"longitude",
								"latitude",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AddRouteStopDTO": {
				"properties": {
					"routeId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"sequence": {
						"type": "number",
						"format": "double"
					},
					"estimatedTime": {
						"type": "string"
					}
				},
				"required": [
					"routeId",
					"name",
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "busapp-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/routes": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "List all transit routes for a college campus.\n\n### Intent & Business Purpose\nFetches bus transit routes, start/end locations, and ordered pickup stops.\n\n### Target Audience & Roles\n- **Allowed Roles:** Authenticated users belonging to the tenant college.",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Optional college UUID filter for Super Admins.",
						"in": "query",
						"name": "collegeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create a new transit route.\n\n### Intent & Business Purpose\nCreates a new transit route with name, start point, end point, and college tenant assignment.\n\n### Request Body\n- `name` *(required string)*: Route name (e.g. `Koteshwor - Balkhu Express`).\n- `startLocation` *(required string)*: Route origin location.\n- `endLocation` *(required string)*: Route destination location.\n- `collegeId` *(required string)*: College UUID.",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Route creation payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRouteDTO",
								"description": "Route creation payload."
							}
						}
					}
				}
			}
		},
		"/api/v1/routes/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get route details and stop list by ID.\n\n### Path Parameters\n- `id` *(required string)*: Route UUID.",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Route UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update route info.\n\n### Path Parameters\n- `id` *(required string)*: Route UUID.",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Route UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update fields.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRouteDTO",
								"description": "Update fields."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Deactivate a transit route.\n\n### Path Parameters\n- `id` *(required string)*: Route UUID.",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Route UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/routes/stops": {
			"post": {
				"operationId": "AddStop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Add a stop to a route.\n\n### Intent & Business Purpose\nAdds an intermediate pickup/dropoff stop with name, sequence order, latitude, longitude, and estimated arrival time.\n\n### Request Body\n- `routeId` *(required string)*: Target route UUID.\n- `stopName` *(required string)*: Name of stop (e.g. `Kalanki Chowk`).\n- `sequence` *(required number)*: Numerical sequence index on the route.\n- `latitude` *(required number)*: Latitude coordinate.\n- `longitude` *(required number)*: Longitude coordinate.\n- `estimatedTime` *(optional string)*: Estimated arrival time (e.g. `07:30 AM`).",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Route stop creation payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddRouteStopDTO",
								"description": "Route stop creation payload."
							}
						}
					}
				}
			}
		},
		"/api/v1/routes/stops/{stopId}": {
			"delete": {
				"operationId": "DeleteStop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Remove a stop from a route.\n\n### Path Parameters\n- `stopId` *(required string)*: Route stop UUID.",
				"tags": [
					"Routes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Stop UUID.",
						"in": "path",
						"name": "stopId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Development server"
		}
	]
}